Tasks
I finish the img for index.

About: take out the first section of about, change position of the
title, take out a block for personal, fit the block in the space.

Contact: take out the form, the the ugly orange icon and implement the
NYC map, plus that, create a style for each change to make it fit in
each screen.